fix numeric issue (floating point division, precision)

+++ src/main/java/de/codengine/krabat/anims/Dinglingerwalk.java
@@ -285,7 +285,7 @@ public class Dinglingerwalk extends Mainanim {
         int scale = getScale((int) yps);
 
         // Zooming - Faktor beruecksichtigen in x - Richtung
-        float horiz_dist = CHORIZ_DIST[anim_pos] - scale / SLOWX;
+        float horiz_dist = CHORIZ_DIST[anim_pos] - (float) scale / SLOWX;
         if (horiz_dist < 1) {
             horiz_dist = 1;
         }
