+++ src/main/java/de/codengine/krabat/anims/Druzina.java
@@ -152,7 +152,7 @@ public class Druzina extends Mainanim {
         int scale = getScale((int) yps);
 
         // Zooming - Faktor beruecksichtigen in y-Richtung
-        float vert_dist = CVERT_DIST[anim_pos] - scale / SLOWY;
+        float vert_dist = CVERT_DIST[anim_pos] - (float) scale / SLOWY;
         if (vert_dist < 1) {
             vert_dist = 1;
             // hier kann noch eine Entscheidungsroutine hin, die je nach Animationsphase
